refactor(form-builder): continue splitting main file into multiple ones
