+++ tools/bazar/presentation/javascripts/form-edit-template/form-edit-template.js
@@ -25,7 +25,9 @@ import custom from './fields/custom.js'
 import tabs from './fields/tabs.js'
 import tabchange from './fields/tabchange.js'
 
-import { defaultMapping } from './fields/commons/attributes.js'
+import { parseWikiTextIntoJsonData, formatJsonDataIntoWikiText } from './yeswiki-syntax-converter.js'
+import { copyMultipleSelectValues, mapFieldsConf } from './form-builder-helper.js'
+import I18nOption from './i18n.js'
 
 const $formBuilderTextInput = $('#form-builder-text')
 let formBuilder
@@ -39,108 +41,15 @@ window.formBuilderFields = {
   listefichesliees, custom, tabs, tabchange
 }
 
-function mapFieldsConf(callback) {
-  return Object.fromEntries(
-    Object.entries(formBuilderFields).map(([name, conf]) => [name, callback(conf)])
-      .filter(([name, conf]) => !!conf)
-  )
-}
-
-// Define an entire group of fields to be added to the stage at a time.
-// Use window to make it available outside of module, so extension can adds their own fields
-window.inputSets = Object.values(formBuilderFields).map((conf) => conf.set).filter((f) => !!f)
-
-// Use window to make it available outside of module, so extension can adds their own fields
-window.yesWikiMapping = mapFieldsConf((conf) => conf.attributesMapping || defaultMapping)
-
-// Mapping betwwen yeswiki field type and standard field implemented by form builder
-// Use window to make it available outside of module, so extension can adds their own fields
-window.yesWikiTypes = {
-  lien_internet: { type: 'url' },
-  lien_internet_bis: { type: 'text', subtype: 'url' },
-  mot_de_passe: { type: 'text', subtype: 'password' },
-  // "nombre": { type: "text", subtype: "tel" },
-  texte: { type: 'text' }, // all other type text subtype (range, text, tel)
-  textelong: { type: 'textarea', subtype: 'textarea' },
-  listedatedeb: { type: 'date' },
-  listedatefin: { type: 'date' },
-  jour: { type: 'date' },
-  map: { type: 'map' },
-  carte_google: { type: 'map' },
-  checkbox: { type: 'checkbox-group', subtype2: 'list' },
-  liste: { type: 'select', subtype2: 'list' },
-  radio: { type: 'radio-group', subtype2: 'list' },
-  checkboxfiche: { type: 'checkbox-group', subtype2: 'form' },
-  listefiche: { type: 'select', subtype2: 'form' },
-  radiofiche: { type: 'radio-group', subtype2: 'form' },
-  fichier: { type: 'file', subtype: 'file' },
-  champs_cache: { type: 'hidden' },
-  listefiches: { type: 'listefichesliees' }
-}
-
-const defaultFieldsName = {
-  textarea: 'bf_description',
-  image: 'bf_image',
-  champs_mail: 'bf_mail',
-  date: 'bf_date_debut_evenement'
-}
-
-const I18nOption = {
-  ar: 'ar-SA',
-  ca: 'ca-ES',
-  cs: 'cs-CZ',
-  da: 'da-DK',
-  de: 'de-DE',
-  el: 'el-GR',
-  en: 'en-US',
-  es: 'es-ES',
-  fa: 'fa-IR',
-  fi: 'fi-FI',
-  fr: 'fr-FR',
-  he: 'he-IL',
-  hu: 'hu-HU',
-  it: 'it-IT',
-  ja: 'ja-JP',
-  my: 'my-MM',
-  nb: 'nb-NO',
-  pl: 'pl-PL',
-  pt: 'pt-BR',
-  qz: 'qz-MM',
-  ro: 'ro-RO',
-  ru: 'ru-RU',
-  sj: 'sl-SL',
-  th: 'th-TH',
-  uk: 'uk-UA',
-  vi: 'vi-VN',
-  zh: 'zh-CN'
-}
-
-function copyMultipleSelectValues(currentField) {
-  const currentId = $(currentField).prop('id')
-  // based on formBuilder/Helpers.js 'incrementId' function
-  const split = currentId.lastIndexOf('-')
-  const clonedFieldNumber = parseInt(currentId.substring(split + 1)) - 1
-  const baseString = currentId.substring(0, split)
-  const clonedId = `${baseString}-${clonedFieldNumber}`
-
-  // find cloned field
-  const clonedField = $(`#${clonedId}`)
-  if (clonedField.length > 0) {
-    // copy multiple select
-    const clonedFieldSelects = $(clonedField).find('select[multiple=true]')
-    clonedFieldSelects.each(function() {
-      const currentSelect = $(currentField).find(`select[multiple=true][name=${$(this).prop('name')}]`)
-      currentSelect.val($(this).val())
-    })
-  }
-}
-
-const typeUserEvents = {}
-Object.keys(formBuilderFields).forEach((field) => {
-  typeUserEvents[field] = { onclone: copyMultipleSelectValues }
-})
-
 function initializeFormbuilder() {
+  // Define an entire group of fields to be added to the stage at a time.
+  const inputSets = Object.values(formBuilderFields).map((conf) => conf.set).filter((f) => !!f)
+
+  const typeUserEvents = {}
+  Object.keys(formBuilderFields).forEach((field) => {
+    typeUserEvents[field] = { onclone: copyMultipleSelectValues }
+  })
+
   // FormBuilder conf
   formBuilder = $('#form-builder-container').formBuilder({
     showActionButtons: false,
@@ -205,10 +114,11 @@ function initializeFormbuilder() {
     }
   })
 
-  // Each 300ms update the text field converting form bulder content into wiki syntax
+  const defaultFieldsName = mapFieldsConf((conf) => conf.defaultIdentifier)
+
   let formBuilderInitialized = false
-  let existingFieldsNames = []; let
-    existingFieldsIds = []
+  let existingFieldsNames = []
+  let existingFieldsIds = []
 
   setInterval(() => {
     if (!formBuilder || !formBuilder.actions || !formBuilder.actions.setData) return
@@ -245,22 +155,23 @@ function initializeFormbuilder() {
       $(this).val(newValue)
     })
 
+    // Update the text field converting form builder content into wiki syntax
     if ($('#form-builder-container').is(':visible')) {
       const formData = formBuilder.actions.getData()
       const wikiText = formatJsonDataIntoWikiText(formData)
       if (wikiText) $formBuilderTextInput.val(wikiText)
     }
 
-    // when selecting between data source lists or forms, we need to populate again the listOfFormId select with the
-    // proper set of options
+    // when selecting between data source lists or forms, we need to populate again the
+    // listOfFormId select with the proper set of options
     $('.radio-group-field, .checkbox-group-field, .select-field')
       .find('select[name=subtype2]:not(.initialized)')
-      .change(function() {
+      .on('change', function() {
         $(this).addClass('initialized')
         const visibleSelect = $(this)
           .closest('.form-field')
           .find('select[name=listeOrFormId]')
-        selectedValue = visibleSelect.val()
+        const selectedValue = visibleSelect.val()
         visibleSelect.empty()
         const optionToAddToSelect = $(this)
           .closest('.form-field')
@@ -291,7 +202,7 @@ function initializeFormbuilder() {
           name = defaultFieldsName[fieldType] || `bf_${fieldType}`
           if (existingFieldsNames.includes(name)) {
             // If name already exist, we add a number (bf_address, bf_address1, bf_address2...)
-            number = 1
+            let number = 1
             while (existingFieldsNames.includes(name + number)) number += 1
             name += number
           }
@@ -315,22 +226,21 @@ function initializeFormbuilder() {
     existingFieldsIds = getFieldsIds()
 
     $('.text-field select[name=subtype]:not(.initialized)').on('change', function() {
-        $(this).addClass('initialized')
-        const $parent = $(this).closest('.form-field')
-        if ($(this).val() == 'range' || $(this).val() == 'number') {
-          $parent.find('.maxlength-wrap label').text(_t('BAZ_FORM_EDIT_MAX_VAL'))
-          $parent.find('.size-wrap label').text(_t('BAZ_FORM_EDIT_MIN_VAL'))
-        } else {
-          $parent.find('.maxlength-wrap label').text(_t('BAZ_FORM_EDIT_MAX_LENGTH'))
-          $parent.find('.size-wrap label').text(_t('BAZ_FORM_EDIT_NB_CHARS'))
-        }
-        if ($(this).val() == 'color') {
-          $parent.find('.maxlength-wrap, .size-wrap').hide()
-        } else {
-          $parent.find('.maxlength-wrap, .size-wrap').show()
-        }
-      })
-      .trigger('change')
+      $(this).addClass('initialized')
+      const $parent = $(this).closest('.form-field')
+      if ($(this).val() == 'range' || $(this).val() == 'number') {
+        $parent.find('.maxlength-wrap label').text(_t('BAZ_FORM_EDIT_MAX_VAL'))
+        $parent.find('.size-wrap label').text(_t('BAZ_FORM_EDIT_MIN_VAL'))
+      } else {
+        $parent.find('.maxlength-wrap label').text(_t('BAZ_FORM_EDIT_MAX_LENGTH'))
+        $parent.find('.size-wrap label').text(_t('BAZ_FORM_EDIT_NB_CHARS'))
+      }
+      if ($(this).val() == 'color') {
+        $parent.find('.maxlength-wrap, .size-wrap').hide()
+      } else {
+        $parent.find('.maxlength-wrap, .size-wrap').show()
+      }
+    }).trigger('change')
 
     // in semantic field, we want to separate value by coma
     $('.fld-semantic').each(function() {
@@ -363,128 +273,9 @@ function getFieldsIds() {
   return result
 }
 
-// Remove accidental br at the end of the labels
-function removeBR(text) {
-  let newValue = text.replace(/(<div><br><\/div>)+$/g, '')
-  // replace multiple '<div><br></div>' when at the end of the value
-  newValue = newValue.replace(/(<br>)+$/g, '')
-  // replace multiple '<br>' when at the end of the value
-  return newValue
-}
-
 function initializeBuilderFromTextInput() {
   const jsonData = parseWikiTextIntoJsonData($formBuilderTextInput.val())
   formBuilder.actions.setData(JSON.stringify(jsonData))
-}
-
-// transform a json object like "{ type: 'texte', name: 'bf_titre', label: 'Nom' .... }"
-// into wiki text like "texte***bf_titre***Nom***255***255*** *** *** ***1***0***"
-function formatJsonDataIntoWikiText(formData) {
-  if (formData.length == 0) return null
-  let wikiText = ''
-
-  for (let i = 0; i < formData.length; i++) {
-    const wikiProps = {}
-    const formElement = formData[i]
-    const mapping = yesWikiMapping[formElement.type]
-
-    for (const type in yesWikiTypes) {
-      if (
-        formElement.type == yesWikiTypes[type].type
-        && (!formElement.subtype
-          || !yesWikiTypes[type].subtype
-          || formElement.subtype == yesWikiTypes[type].subtype)
-        && (!formElement.subtype2
-          || formElement.subtype2 == yesWikiTypes[type].subtype2)
-      ) {
-        wikiProps[0] = type
-        break
-      }
-    }
-    // for non mapped fields, we just keep the form type
-    if (!wikiProps[0]) wikiProps[0] = formElement.type
-
-    // fix for url field which can be build with textField or urlField
-    if (wikiProps[0]) wikiProps[0] = wikiProps[0].replace('_bis', '')
-
-    for (const key in mapping) {
-      const property = mapping[key]
-      if (property != 'type') {
-        let value = formElement[property]
-        if (['required', 'access'].indexOf(property) > -1) value = value ? '1' : '0'
-        if (property == 'label') {
-          wikiProps[key] = removeBR(value).replace(/\n$/gm, '')
-        } else {
-          wikiProps[key] = value
-        }
-      }
-    }
-
-    const maxProp = Math.max.apply(Math, Object.keys(wikiProps))
-    for (let j = 0; j <= maxProp; j++) {
-      wikiText += wikiProps[j] || ' '
-      wikiText += '***'
-    }
-    wikiText += '\n'
-  }
-  return wikiText
-}
-
-// transform text with wiki text like "texte***bf_titre***Nom***255***255*** *** *** ***1***0***"
-// into a json object "{ type: 'texte', name: 'bf_titre', label: 'Nom' .... }"
-function parseWikiTextIntoJsonData(text) {
-  const result = []
-  var text = text.trim()
-  const textFields = text.split('\n')
-  for (let i = 0; i < textFields.length; i++) {
-    const textField = textFields[i]
-    const fieldValues = textField.split('***')
-    const fieldObject = {}
-    if (fieldValues.length > 1) {
-      const wikiType = fieldValues[0]
-      let fieldType = wikiType in yesWikiTypes ? yesWikiTypes[wikiType].type : wikiType
-      // check that the fieldType really exists in our form builder
-      if (!(fieldType in yesWikiMapping)) fieldType = 'custom'
-
-      const mapping = yesWikiMapping[fieldType]
-
-      fieldObject.type = fieldType
-      fieldObject.subtype = wikiType in yesWikiTypes ? yesWikiTypes[wikiType].subtype : ''
-      fieldObject.subtype2 = wikiType in yesWikiTypes ? yesWikiTypes[wikiType].subtype2 : ''
-      const start = fieldType == 'custom' ? 0 : 1
-      for (let j = start; j < fieldValues.length; j++) {
-        let value = fieldValues[j]
-        const field = mapping && j in mapping ? mapping[j] : j
-        if (field == 'required') value = value == '1'
-        if (field) {
-          if (field == 'read' || field == 'write' || field == 'comment') {
-            fieldObject[field] = (value.trim() === '')
-              ? (
-                field == 'comment'
-                  ? [' + ']
-                  : [' * ']
-              )
-              : value.split(',').map((e) => ((['+', '*', '%'].includes(e.trim())) ? ` ${e.trim()} ` : e))
-          } else if (field == 'seeEmailAcls') {
-            fieldObject[field] = (value.trim() === '')
-              ? ' % ' // if not define in tempalte, choose owner and admins
-              : value.split(',').map((e) => ((['+', '*', '%'].includes(e.trim())) ? ` ${e.trim()} ` : e))
-          } else {
-            fieldObject[field] = value
-          }
-        }
-      }
-      if (!fieldObject.label) {
-        fieldObject.label = wikiType
-        for (let k = 0; k < fields.length; k++) if (fields[k].name == wikiType) fieldObject.label = fields[k].label
-      }
-      result.push(fieldObject)
-    }
-  }
-  if (wiki.isDebugEnabled) {
-    console.log('parse result', result)
-  }
-  return result
 }
 
 $('a[href="#formbuilder"]').on('click', (event) => {
