+++ tools/bazar/presentation/javascripts/form-edit-template/i18n.js
@@ -1,0 +1,30 @@
+// list of available locales for formBuilder 
+export default {
+  ar: 'ar-SA',
+  ca: 'ca-ES',
+  cs: 'cs-CZ',
+  da: 'da-DK',
+  de: 'de-DE',
+  el: 'el-GR',
+  en: 'en-US',
+  es: 'es-ES',
+  fa: 'fa-IR',
+  fi: 'fi-FI',
+  fr: 'fr-FR',
+  he: 'he-IL',
+  hu: 'hu-HU',
+  it: 'it-IT',
+  ja: 'ja-JP',
+  my: 'my-MM',
+  nb: 'nb-NO',
+  pl: 'pl-PL',
+  pt: 'pt-BR',
+  qz: 'qz-MM',
+  ro: 'ro-RO',
+  ru: 'ru-RU',
+  sj: 'sl-SL',
+  th: 'th-TH',
+  uk: 'uk-UA',
+  vi: 'vi-VN',
+  zh: 'zh-CN'
+}
